==> Code/01_API_Getting Data.ipynb <==
import requests
import time 
import pandas as pd

url = 'https://www.reddit.com/r/disneyparks.json'

headers = {'user-agent': 'Jessica 0.1'}

res = requests.get(url,headers=headers)

res.status_code

json = res.json()

sorted(json.keys())

json['data'].keys()

len(json['data']['children'])

disney_df = pd.DataFrame(json['data']['children']) 

disney_df.head(3)

json['data']['after']

[post['data']['name'] for post in json['data']['children']]

param = {'after': 't3_1bfjjvr'}

requests.get(url,params=param,headers=headers)

### Key Exploration to filter out the important keys

json['data']['children'][1]['data'].keys()

info = ['name','subreddit','title','selftext','upvote_ratio','num_comments','created_utc']
for item in info:
    print(item,':',json['data']['children'][1]['data'][item])

### Data Collection

# Only need these information 'name','subreddit','title','selftext','upvote_ratio','num_comments','created_utc'

def get_subreddit(url):

    posts_data = []
    after = None
    for i in range(40):
            if after ==None:
                params={}
            else:
                params = {'after':after}
            url=url
            res = requests.get(url, params=params, headers=headers)
            if res.status_code == 200:
                the_json = res.json()
                for post in the_json['data']['children']:
                    post_data = {
                        'name': post['data']['name'],
                        'subreddit': post['data']['subreddit'],
                        'title': post['data']['title'],
                        'selftext': post['data']['selftext'],
                        'upvote_ratio': post['data']['upvote_ratio'],
                        'num_comments': post['data']['num_comments'],
                        'created_utc': post['data']['created_utc']
                    }
                    posts_data.append(post_data)
                after = the_json['data']['after']
            else:
                print(res.status_code)
                break
            time.sleep(1)
        
    df = pd.DataFrame(posts_data)
    return df

disneyland_url = 'https://www.reddit.com/r/disneyparks/new.json'
disneyland_df = get_subreddit(disneyland_url)
disneyland_df.tail(3)

disneyland_df.shape

disneyland_df['name'].nunique()

universalstudios_url = 'https://www.reddit.com/r/universalstudios/new.json'
universalstudios_df = get_subreddit(universalstudios_url)
universalstudios_df.head(3)

universalstudios_df.shape

universalstudios_df['name'].nunique()

disneyland_df.to_csv('../data/disneyland.csv')
universalstudios_df.to_csv('../data/universalstudios.csv')
