==> Code/04_All_ModelTraining.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

from sklearn.pipeline import Pipeline
from sklearn.svm import SVC

from sklearn.metrics import fbeta_score, make_scorer
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import accuracy_score,f1_score
from sklearn.preprocessing import StandardScaler,PowerTransformer,Normalizer

from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier, VotingClassifier
from sklearn.metrics import confusion_matrix, ConfusionMatrixDisplay, f1_score, roc_auc_score, RocCurveDisplay, roc_curve
from sklearn.ensemble import RandomForestClassifier


from google.colab import drive
drive.mount('/content/drive')

df = pd.read_csv('../content/drive/MyDrive/project_3/Data/cleaned_combined_df.csv')

df.head(3)

### train_test_split

X = df['title_selftext']
y = df['subreddit']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, stratify=y, random_state=42)

def analyse_model_performance(model,gs_params):

    pipe_gs = GridSearchCV(model, gs_params, cv=5, n_jobs=-1)

    pipe_gs.fit(X_train,y_train)


    # Get predictions
    preds = pipe_gs.predict(X_test)

    # Save confusion matrix values
    tn, fp, fn, tp = confusion_matrix(y_test, preds).ravel()

    best_params = pipe_gs.best_params_

    dict_scores = {
        'train_accuracy_score': accuracy_score(y_train, pipe_gs.predict(X_train)),
        'test_accuracy_score': accuracy_score(y_test, preds),
        'sensitivity': tp / (tp + fn),
        'specificity': tn / (tn + fp),
        'f1': f1_score(y_test, preds),
        'best Params':best_params

    }
    return dict_scores

# All Model training Combination

scores = []
scores_dict = {}

### Logistic Regression

gs_params = {
    'tvec__ngram_range': [(1, 2)],
    'tvec__max_features': [50000, 53500],
    'Logr__penalty': ['l1', 'l2'],
    'Logr__C': np.logspace(-2, 2, 50),
    'Logr__solver': ['liblinear', 'lbfgs'],
    'Logr__random_state': [123]
}

pipe_logr = Pipeline([
    ('tvec', TfidfVectorizer()),
    ('Logr', LogisticRegression())
])

scores_dict = analyse_model_performance(pipe_logr,gs_params)
scores_dict['model'] ="TF-IDF + Logistic Regression"
scores.append(scores_dict)

### Naive Bayes

gs_params = {
    'tvec__ngram_range': [(1, 2)],
    'tvec__max_features': [50000, 53500],
    'NB__alpha': [0.1, 1.0, 10.0]
}

pipe_NB = Pipeline([
    ('tvec', TfidfVectorizer()),
     ('NB', MultinomialNB())
     ])

scores_dict = analyse_model_performance(pipe_NB,gs_params)
scores_dict['model'] ="TF-IDF + Multinomial NB"
scores.append(scores_dict)

### Support Vector Classifier

gs_params = {
    'tvec__ngram_range': [(1,2)],
     'tvec__max_features': [50000, 53500],
    'svc__kernel': ['linear', 'rbf'],
    'svc__C': [0.1, 1, 10],
    'svc__gamma': ['scale', 'auto']

}


pipe_SVC = Pipeline([
    ('tvec', TfidfVectorizer()),
    ('svc', SVC())
])

scores_dict = analyse_model_performance(pipe_SVC,gs_params)
scores_dict['model'] ="TF-IDF + SVC"
scores.append(scores_dict)

### Ensemble Methods

### Random Forest

gs_params = {
    'tvec__ngram_range': [(1, 2)],
    'tvec__max_features': [50000, 53500],
    'rf__n_estimators': [100, 200, 300],
    'rf__max_depth': [None, 10, 20],
    'rf__min_samples_split': [2, 5, 10],
    'rf__min_samples_leaf': [1, 2, 4]
}


pipe_SVC = Pipeline([
    ('tvec', TfidfVectorizer()),
    ('rf', RandomForestClassifier())
])

scores_dict = analyse_model_performance(pipe_SVC,gs_params)
scores_dict['model'] ="TF-IDF + Random Forest"
scores.append(scores_dict)

gs_params = {
    'tvec__ngram_range': [(1, 2)],
    'tvec__max_features': [50000, 53500],
    'ensemble__logr__penalty': ['l2'],
    'ensemble__logr__C': [18.420699693267146],
    'ensemble__logr__solver': ['liblinear'],
    'ensemble__logr__random_state': [123],
    'ensemble__nb__alpha': [0.1],
    'ensemble__svc__gamma': ['scale'],
    'ensemble__svc__kernel': ['rbf'],
    'ensemble__svc__C': [1]
}

pipe_ensemble = Pipeline([
    ('tvec', TfidfVectorizer()),
    ('ensemble', VotingClassifier([
        ('logr', LogisticRegression()),
        ('svc', SVC()),
        ('nb', MultinomialNB())
    ]))
])

scores_dict = analyse_model_performance(pipe_ensemble,gs_params)
scores_dict['model'] ="TF-IDF + Logr + SVC + NB"
scores.append(scores_dict)

df_scores = pd.DataFrame(scores)
df_scores = df_scores.sort_values(by='f1', ascending=False)
df_scores = df_scores.reset_index(drop=True)

model_column = df_scores.pop('model') # Move model from last column to the first
df_scores.insert(0,'model',model_column)
df_scores

Based on the above tables, ensemble methods consisting of the top 3 models performed the best based on F1 score. However, it is still suffering from overfitting issue where the train accuracy score remain 1.

https://github.com/facebookresearch/fastText/tree/main/python
