==> Code/05_FastText.ipynb <==
! pip install fasttext

from google.colab import drive
drive.mount('/content/drive')

import fasttext
from sklearn.model_selection import GridSearchCV
import pandas as pd

### Fast Text

train_df = '../content/drive/MyDrive/project_3/Data/cleaned_combined_train_df'
valid_df = '../content/drive/MyDrive/project_3/Data/cleaned_combined_valid_df'
test_df = '../content/drive/MyDrive/project_3/Data/cleaned_combined_test_df'

model = fasttext.train_supervised(input= train_df , autotuneValidationFile= valid_df)

score_test = model.test(test_df)
score_train = model.test(train_df)

def return_f1_score(score):

  dict_scores = {
    'precision': score[1],
    'recall': score[2],
    'F1': 2 * ((score[1]* score[2]) /(score[1] + score[2]))
  }

  return dict_scores

# Test score
return_f1_score(score_test)

#Train dataset
return_f1_score(score_train)

Over fitting issue

df = pd.read_csv('../content/drive/MyDrive/project_3/Data/cleaned_combined_df.csv')

# Test
test_df_1 = df.sample(frac=0.20, random_state=42)

# Train
train_df_1 = df.drop(test_df_1.index)

train_df_1.to_csv('../content/drive/MyDrive/project_3/Data/cleaned_combined_train_df_1',columns=['subreddit_title_selftext'],index=False,header=False)
test_df_1.to_csv('../content/drive/MyDrive/project_3/Data/cleaned_combined_test_df_1',columns=['subreddit_title_selftext'],index=False,header=False)

train_df_1 = '../content/drive/MyDrive/project_3/Data/cleaned_combined_train_df_1'
test_df_1 = '../content/drive/MyDrive/project_3/Data/cleaned_combined_test_df_1'

model_2 = fasttext.train_supervised(input= train_df_1 , lr=1.0, epoch=55, wordNgrams = 2, bucket=200000, dim=100, loss='hs' , minCount=5)

score_test = model_2.test(test_df_1)
score_train = model_2.test(train_df_1)

# Test score
return_f1_score(score_test)

#Train dataset
return_f1_score(score_train)

The F1 score on the training dataset shows perfect performance (1.0) for both automatic hyperparameter tuning and random search methods. However, on the testing dataset, the F1 score for automatic hyperparameter tuning is lower than that for random search by 0.01. This discrepancy may be attributed to several factors.

Firstly, the automatic hyperparameter tuning selects the best parameters based on cross-validation results on a validation dataset, while random search directly evaluates on the test dataset. If there are significant differences in the distributions between the training and test datasets, the parameters optimized for the training data might not generalize well to unseen data.

Secondly, insufficient data for training and testing could result in both precision and recall scores achieving 1.0.

Thirdly, the two subreddits  are highly distinguishable with unique bigram. As a result, there might be numerous hyperparameter combinations capable of achieving perfect accuracy or F1 scores, leading to not getting the truth hyperparameter that would also perform well on test dataset.

https://github.com/facebookresearch/fastText/tree/main/python
